Chat floating action button added.
A floating action button added at the buttom, that is for chat. This
button is added on the CHATS page, and for that a new directory is
creaetd named Pages which will consists of all the different pages, and
in that directory a new file named ChatPage.dart has this button.